Setup the join page into the Chat Application
created a form that allow a user to join a room by entering a username and the room name to join.

implemented socket.io room. fetch username and room name from the query string using the qs library, emit join event with the username and room name grab from the query string, created an event listener to listen to the join event. used the socket.joint() method to join a room. used the to() method to broacast message to a specific room

Closes issues #19 and #20